Add EntityConfig deletable() option

+++ resources/views/cms/partials/list/delete-button.blade.php
@@ -1,4 +1,4 @@
-@if(check_ability('delete', $category->key(), $entity->key(), $instance->id))
+@if($entity->deletable() && check_ability('delete', $category->key(), $entity->key(), $instance->id))
     <li>
         {!! Form::open(["route"=>["sharp.cms.destroy", $category->key(), $entity->key(), $instance->id], "method"=>"DELETE", "id"=>"sharpdelete".$instance->id]) !!}
         {!! Form::close() !!}
